Added: drone logic, HUD menu, Readme file

+++ Assets/Scripts/Data/GamePrefs/Config/IGamePrefsConfig.cs
@@ -2,7 +2,7 @@
 {
 	public interface IGamePrefsConfig
 	{
-		(float min, float max) DroneCount
+		(int min, int max) DroneCount
 		{
 			get;
 		}
